refactor: rename util and type directories to utils and types for consistency
