+++ src/components/sign/hooks/useInputName.ts
@@ -1,6 +1,6 @@
 'use client';
 
-import { HandleInputEvent } from '@/type';
+import { HandleInputEvent } from '@/types';
 import { useSignStates } from './useSignStates';
 
 function useInputName({ setUsername }: ReturnType<typeof useSignStates>) {
